wl_build: Also use --with-x-locale-root for xkbcommon
libxkbcommon by default will look under $prefix/share/X11 for locale
files.  Direct it to use the system locale files in /usr/share/X11
instead.

This shows up as a problem when using the 'include "%L"' directive in
compose files (e.g. .XCompose) to pull in the locale-specific Compose
file -- xkbcommon looks in the wrong place and returns an error such as:

  xkbcommon: ERROR: ~/.XCompose:4:9: failed to expand %L to the locale Compose file
  xkbcommon: ERROR: ~/.XCompose:4:12: unterminated string literal

Signed-off-by: Bryce Harrington <bryce@osg.samsung.com>